pyproject.toml
[project]
name = "boletimocorrencias-t3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi[standard]>=0.128.0",
    "pydantic-settings>=2.12.0",
    "pymongo[srv]>=4.16.0",
    "python-dotenv>=1.2.1",
]
